backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
sqlmodel==0.0.22
groq==0.13.1
PyJWT==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo[srv]==4.6.3

